refactor of help-me-carry challenge, works in simulation
