+++ robocup_knowledge/src/robocup_knowledge/environments/robotics_testlabs/challenge_help_me_carry.py
@@ -1,31 +1,28 @@
-commands = {
-    'follow':               ["follow"],    # 'Robot, follow me'
-    'follow_or_remember':   ["follow",     # 'Robot, follow me'
-                             "remember",   # 'Remember, this location is the car'
-                             "car",
-                             "stop"],
-    'carry':                ["carry",
-                             "bag",
-                             "bring"]}     # 'Carry this bag'
-
-# add or get from ~/ros/kinetic/system/src/ed_object_models/models/robotics_testlabs/model.yaml
-waypoints = {
-    'bedroom':              {'id':      "explore6",
-                             'radius':  1},
-    'livingroom':              {'id':   "operator_pose",
-                             'radius':  1},
-    }
-
-waypoint_car = {'id':      "car",
-                'radius':  1}
-
-default_item = "coke"
-default_place = "cabinet"
-default_area = "on_top_of"
-
-starting_point = "help_me_carry_starting_point"
+starting_point = 'help_me_carry_starting_point'
 rotation = 0
 
 time_out_seconds = 60.0
 time_out_seconds_door = 120.0
 
+commands = {
+    'follow':               ['follow',              # 'Robot, follow me'
+                             'follow me'],          
+    'remember':             ['remember location',   # 'Remember, this location is the car'
+                             'here is the car',     
+                             'stop following',      # 'Stop following me'
+                             'stop following me']}     
+                             
+# add or get from ~/ros/kinetic/system/src/ed_object_models/models/robotics_testlabs/model.yaml
+waypoints = {
+    'dinner_table':         {'id':   'dinner_table',
+                             'radius':  1},
+    'kitchen':              {'id':   'bookcase_waypoint',
+                             'radius':  1},
+    }
+
+waypoint_car = {'id':      'car',
+                'radius':  1}
+
+default_item = 'coke'
+default_place = 'cabinet'
+default_area = 'on_top_of'
